move tests for new PV functions to separate file
